Move updatable auto extractor types to /models
